Update README
I've removed the direct links to el6 rpms, which might not work for everyone. This plugin is already in the official repos for 1.6 and 1.7 so they can just use that. 

Rbovirt is a dependency so there's no need to tell people explicitely to install it. I've added a small section about installing from the gem directly too, and a foreman/plugin compatibility versions table too.